refactor: feature envy
ensuring srp, renamed also method
